requirements.txt
pysam >= 0.7.7
pyVCF >= 0.6.7
biopython >= 1.61
